requirements.txt
Jetson.GPIO
logging
time
socket
adafruit-io
requests